Integrate the plotting into the package
Might as well. This does bring in the heavy GLMakie dependency now,
but that's okay. I still can't figure out the recipe system when it's
not just the plot that needs to be defined, but also a specific format
for the figure and axis, etc. Apparently their recipe system is still
subject to change as well (maybe I'm wrong about this), and MakieCore
doesn't bring in Colorbar, etc. either so maybe it's best to just do
this for now.

It's more convenient to have the plotting upfront and accessible while
controlling the dependency's version rather than defining a specific
recipe and requiring the user manage it. Whenever I figure out recipes
I'll just change it.

Also, I could just use Plots.jl or another lighter package since right
now there's no interactivity required, but I like Makie and since I'm
more familiar with manipulating elements in it I'm just going to go with
that; it's more robust anyway from my experience and nowadays with newer
versions of Julia and Makie loading it is very fast.