Use a different modulus transform
This transform has a smaller effective slope for values < 1 (L1 ~ arctan
has a secant line slope of ~0.6 from the origin at x = 0.5, while L2 has
a secant of 0.4) which covers half of the Lightness range, but
compresses more aggressively as it binds tighter more rapidly thus
encoding the large range of possible values at the top end into a
smaller range of values close to 1; for these values the actual
brightness doesn't matter much since perceptually we're less sensitive
to changes between very bright tones, so we're unlikely to notice
differences for very large moduli, but the extent of bright values is
expanded on the plot with this new definition.

Since we're more sensitive to black level differentials with a response
which is approximately logarithmic in relation (Fechner's Law) or has a
cubic root to square root dependence (Steven's Law) then the gradient is
rolled off quite nicely in a visual sense near the roots and for very
small values.

Anyway, to quantify the differences between the two functions in a rough
sense: for x = [3.0, 5.0, 10.0]:
L1(x) ≈ [0.80, 0.87, 0.94]; L2(x) ≈ [0.90, 0.96, 0.99].

The transform is the bijective conformal mapping from the complex plane
onto the Riemann sphere with a radius of 1/2, centered at 1/2.

The inverse stereographic projection onto the top hemisphere of the
Riemann sphere with radius 1, centered at 0 with the complex plane
slicing through its equator yields:
ζ = (|z|² - 1) / (|z|² + 1), but this doesn't map onto zero for z = 0.

Instead, for the shifted and scaled sphere on top of and touching the
complex plane we have:
|w|² = (1 - ζ)ζ;
(1 - ζ)² * |z|² = (1 - ζ)ζ;
which gives ζ = |z|² / (|z|² + 1) where ζ is the real-valued Z point
on the sphere for (x, y, ζ) ∈ ℝ with coordinates expressed as
(z, ζ) where z ∈ ℂ.

So it's essentially the z ↦ Z mapping under ℝ² → S² from the plane onto
the sphere given z = w / (1 - ζ) under the forward projection onto the
plane for S² → ℝ² with |w|² + (ζ - 0.5)² = 0.25.

Anyway, it looks better, so I'm going to go with that.