First attempt at discontinuous plotting
I will likely leave in the phase approach, but for the modulus I think
a direct contour solution is more appropriate so that will be the next
commit. The goal is to highlight |z| = 2^n transitions.

Here the scaling could be applied to either:

* the modulus
* the Lightness
* the RGB matrix

or instead of applying the Lightness r2 /(r2 + 1) transform we could
instantiate a matrix of 0.5 values for L and then scale that. I've
decided to scale the Lightness here, but as I said this will be deleted
in the next commit.