Add contours
So not entirely ideal as the way Makie plots the phase contours adds
extra lines at around ±180°, but maybe I'm calculating things
incorrectly and I need to employ some trick in order to avoid this.
It's not a big deal though, the plot is still correct and looks good
regardless.

There's a bit of added complexity due to a micro-optimization in order
to avoid recomputing the modulus matrix: we use the modulus for the
contours, but also in computing the Lightness; this way we get to have
two methods for `complex_colors` - one for the complex array, one for an
input set of modulus and phase matrices.

Note that `image!` must be called _before_ the `contour!` calls so that
the contours will get drawn over the image.

Also, `septaphase` was chosen for several reasons, and is justifed as
follows:

* 'septum' is Latin for 'partition', 'wall', 'enclosure', 'fence'
  so it seems entirely appropriate;

* 'septa-' is the prefix for seven (from the Greek 'hepta'), and even
  though we have six phase partitions this is counting the n + 1
  positions required for a full revolution in much the same way an
  octave in music theory refers to the interval of 7 notes in a diatonic
  scale plus the tonic note which is the next pitch in the same pitch
  class as the first; so in essence eight are required for completion
  and so analogously seven transitions are required for resolution under
  our system. Namely this refers to the phases n * π/3, n = 0,1,...,6.