Simplify
Continued from (part1: e4b3084, part 2: 85faa56).

This is effectively part 3 (likely the finalizing step) of the refactor
which was necessary to make the plotting interactive to some degree.

It didn't make sense to squash the commits, as the division makes it
easier to understand the changes, but at the same time the three commits
don't really stand alone individually and are considered part of a
package of changes with the same objective in mind; part 1 laid the
foundation and prepared the addition of dynamic elements, part 2 added
the bulk of the interactivity, and this commit finalizes things by
refactoring again and simplifying things by reducing unnecessary
complexity. The previous two commits don't have commit messages because
it was easier in this case to finish everything and then explain it in
one go.

Note that part 1 also added a complex_plot method for input tuples
(which were previously considered part of a Union with AbstractVectors),
where now you can specify the Re and Im limits without having to pass
the actual vectors or ranges. This is part of the interactivity as the
DataInspector will index these vectors to determine the position on the
plot and so it was necessary to construct them if only two-tuples were
being passed.

Now, onto the actual changes:

* The contour plots (for both modulus and phase) are now toggleable;

* `septaphase` now has two distinct presentations:
  one by rounding, and one by thresholding;

* DataInspector is now employed and returns all four of the relevant
  dimensions (Re(z), Im(z), abs(s), Arg(s)).

The inspector is especially useful since the range ticks on the plot are
static and the complex magnitude can vary rapidly and have a high
dynamic range, and so this allows for more detailed information querying
and granular resolution through indexing the relevant arrays.

Note that at the moment of writing this commit this inspector is bugged
with RGB arrays and so requires the fix in Makie PR 3468.

Anyway, the keyword arguments have been done away with and the options
are controlled with the interactive elements (septaphase with a slider).

Also, the complex_color method accepting modulus and phase matrices has
been removed for the sake of simplicity; although this does require
extra computation if you already have the modulus and phase arrays
separated it's really not an issue and it's also somewhat arbitrary to
accept that and not the modulus squared instead (especially since that's
what's used to actually compute the image in the end anyway).

Many of the "micro-optimizations" have been removed in order to favour
maintainability; it's debatable whether the efficiency was better
in practice anyway as many of the redundant computations are required
now in any case and it makes more sense to pre-compute the arrays and
have them on stand-by for the interactivity. For example, having a
polar2HSL method which accepts a degree-specified phase adds unnecessary
complexity, has no distinct dispatch distinguishing it from a radian-
specified phase, and this phase needs to be further operated on anyway
to prepare it into a proper array for the Hue component; this method was
never added, but I'm using it to illustrate by way of example.

A final note on the use of ldexp insted of 2.0^n: this isn't easier to
read, but is in fact simpler in implementation in the sense that it's
much faster (by around an order of magnitude); this isn't too important
right now since we're only generating a 12-element vector, but it's
still simply more efficient and Julia requires either the base or the
exponent to be a floating point number if the exponent is negative so
this makes things more explicit (since ldexp methods only accept
(float, integer) type tuples).

P.S. deciding on GLMakie was a good idea even though the program didn't
begin in an interactive state. Had I gone with another plotting package
I would've had to rewrite quite a bit of it here whereas Makie made it
natural to extend the program's functionality; the package is actually
useful now as a result.