Specify the angular offset in degrees
This avoids irrational number approximations as the offset can be set as
an integer.

The Oklch angle is now also guaranteed to reside within the canonical
domain. Colors.jl makes the proper conversions since I believe they
convert to Cartesian coordinates first, which is what the colormap
calculations make use of here, but our Septaphase methods require
adherence to a strict domain for the rounding techniques.

Also, the saturation's element type is no longer specified as it's
unnecessary.