Pull `Colors` from `Sagnac:master`, `IntervalSets` from `Sagnac:range`
Julia 1.11 makes it easy to do this using the sources section so instead
of waiting until things get merged and released this allows the package
immediate access to the features; this can easily be reversed anyway and
once things get updated this will again depend on the registered
packages.

Colors had to be forked here in order to specify a different version
number due to compatibility requirements with Makie.