Adjust the interval range length accordingly for non-unit aspect ratios
Also the figure is now appropriately resized on the first call yielding
a better initial presentation under wide aspects.

The base length value might need to be adjusted and the scaled values
bounded in order to limit the size of the arrays, but for now this is
reasonable.

For very wide aspects the window size should probably also be limited.