Depend upon the new version of `Colors`
IntervalSets will be left in `sources` for ease of reproducing the
environment in case there's a desire to use the other range methods, but
they aren't strictly necessary to use the package.

ComplexColor does export OpenInterval and the interval string macro, but
these aren't necessary to plot since support for the other half-open and
closed intervals is there.

This merge will be pending a Makie release, but their Colors compat has
already been bumped in master.