Add modulus and phase plot switches
Sliders had to be used instead of toggles because I couldn't find a way
to reset the toggle buttons programmatically (toggle.active only sets
the Observable, but not the visual toggle). That's fine as visually this
is more uniform / consistent with the septaphase slider.

There are a few quirks, however, such as the plot switching back to the
septaphase plot on different slider value click 1 (the slider dragging
attribute feature has yet to be merged in Makie so I can't use that to
yield) and on different slider value click 2 the plot will switch, but
not the slider.

Also, the slider widths are now baked in with literals as newer versions
of Makie use `Auto` for the toggle width defaults, so I just used the
default from their manual and scaled it from there.