Don't draw phase contours for constant phase maps
I overlooked this as it should have been made no-op in commit a5f09b2.

Basically plotting either a pure phase or a phase front will cause Makie
to complain and throw an interpolation error on drawing contours since
all the values are practically the same so this needs to be guarded
against.