Apply a non-linearity to the real and imaginary parts
The direct implementation was bugged since the arrays were being color
converted rather than directly imaged and so values were being clipped.

A standard logistic is now applied before conversion; the current
scaling might be too aggressive of a squash for some functions, while
causing a loss of contrast for others, however, so it might be better to
go with standard affine scaling in the long run instead, but for now
this is good.